ADCM-1042 separated the logic between component and service
