+++ web/package.json
@@ -12,8 +12,8 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.0.2",
-    "@angular/cdk": "^8.0.1",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
     "@angular/common": "8.2.14",
     "@angular/compiler": "8.2.14",
     "@angular/core": "8.2.14",
@@ -22,35 +22,35 @@
     "@angular/platform-browser": "8.2.14",
     "@angular/platform-browser-dynamic": "8.2.14",
     "@angular/router": "8.2.14",
-    "@ngrx/effects": "^8.0.1",
-    "@ngrx/entity": "^8.0.1",
-    "@ngrx/router-store": "^8.0.1",
-    "@ngrx/store": "^8.0.1",
-    "@ngrx/store-devtools": "^8.0.1",
+    "@ngrx/effects": "^8.6.0",
+    "@ngrx/entity": "^8.6.0",
+    "@ngrx/router-store": "^8.6.0",
+    "@ngrx/store": "^8.6.0",
+    "@ngrx/store-devtools": "^8.6.0",
     "hammerjs": "^2.0.8",
-    "rxjs": "^6.5.2",
+    "rxjs": "^6.5.4",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.20",
-    "@angular/cli": "8.3.21",
+    "@angular-devkit/build-angular": "~0.803.22",
+    "@angular/cli": "^8.3.22",
     "@angular/compiler-cli": "8.2.14",
     "@angular/language-service": "8.2.14",
-    "@ngrx/schematics": "^8.0.1",
+    "@ngrx/schematics": "^8.6.0",
     "@types/jasmine": "~3.5.0",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~13.1.1",
-    "codelyzer": "^5.0.1",
-    "eslint": "^6.7.2",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "~13.1.5",
+    "codelyzer": "^5.2.1",
+    "eslint": "^6.8.0",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.4.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "^2.0.4",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "node-sass": "^4.12.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "~3.0.3",
+    "karma-jasmine-html-reporter": "^1.5.1",
+    "node-sass": "^4.13.0",
     "protractor": "~5.4.2",
     "ts-node": "~8.5.4",
     "tslint": "~5.20.1",
