+++ web/src/app/shared/components/hostmap/services2hosts/service-host.component.ts
@@ -11,16 +11,16 @@
 // limitations under the License.
 import { animate, state, style, transition, trigger } from '@angular/animations';
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { FormGroup } from '@angular/forms';
 import { ChannelService } from '@app/core';
-import { ApiService } from '@app/core/api';
 import { EventMessage, SocketState } from '@app/core/store';
-import { Component as adcmComponent, Host, IActionParameter } from '@app/core/types';
+import { IActionParameter } from '@app/core/types';
 import { Store } from '@ngrx/store';
-import { Subject, throwError } from 'rxjs';
-import { catchError, take, tap } from 'rxjs/operators';
+import { tap } from 'rxjs/operators';
 
 import { SocketListener } from '../../../directives/base.directive';
-import { CompTile, HostTile, Post, StatePost, Stream, Tile } from '../types';
+import { TakeService } from '../take.service';
+import { CompTile, HostTile, Post } from '../types';
 
 @Component({
   selector: 'app-service-host',
@@ -36,10 +36,16 @@ import { CompTile, HostTile, Post, StatePost, Stream, Tile } from '../types';
   ]
 })
 export class ServiceHostComponent extends SocketListener implements OnInit {
-  countConstraint = 0;
   showSpinner = false;
   showPopup = false;
-  all$ = new Subject<Tile[]>();
+
+  serviceComponents: CompTile[];
+  hosts: HostTile[];
+
+  /**
+   * Validation only
+   */
+  formGroup = new FormGroup({});
 
   @Input()
   cluster: { id: number; hostcomponent: string };
@@ -61,25 +67,17 @@ export class ServiceHostComponent extends SocketListener implements OnInit {
 
   @Output() saveResult = new EventEmitter<Post[]>();
 
-  statePost = new StatePost();
-  loadPost = new StatePost();
-  sourceMap = new Map<string, Tile[]>([
-    ['host', []],
-    ['compo', []]
-  ]);
-  stream = new Stream();
-
   saveFlag = false;
   initFlag = false;
 
   scrollEventData: { direct: 1 | -1 | 0 };
 
-  constructor(private api: ApiService, private channel: ChannelService, socket: Store<SocketState>) {
+  constructor(public service: TakeService, private channel: ChannelService, socket: Store<SocketState>) {
     super(socket);
   }
 
   public get noValid() {
-    return !!this.countConstraint || !this.statePost.data.length;
+    return !!this.service.countConstraint || !this.service.statePost.data.length;
   }
 
   ngOnInit() {
@@ -101,7 +99,7 @@ export class ServiceHostComponent extends SocketListener implements OnInit {
       ((m.event === 'add' || m.event === 'remove') && m.object.details.type === 'cluster' && +m.object.details.value === this.cluster.id)
     ) {
       this.clearRelations();
-      this.checkConstraints();
+      this.service.checkConstraints();
       this.init();
       this.initFlag = true;
     }
@@ -110,226 +108,51 @@ export class ServiceHostComponent extends SocketListener implements OnInit {
   init() {
     if (this.initFlag) return;
     this.initFlag = true;
-
-    this.sourceMap.set('host', []);
-    this.sourceMap.set('compo', []);
-    this.statePost.clear();
-
-    const constraint = (c: number[] | string[]) => {
-      if (c && c[0] && c[0] === '+') c = [this.sourceMap.get('host').length];
-      return c;
-    };
-
-    const getActions = (c: adcmComponent) => {
-      if (this.actionParameters) return this.actionParameters.filter(a => a.service === c.service_name && a.component === c.name).map(b => b.action);
-    };
-
-    this.api
-      .get<{ component: adcmComponent[]; host: Host[]; hc: Post[] }>(this.cluster.hostcomponent)
+    this.service
+      .initSource(this.cluster.hostcomponent, this.actionParameters)
       .pipe(
         tap(a => {
-          if (a.host)
-            this.sourceMap.set(
-              'host',
-              a.host.map(h => new HostTile(h.id, h.fqdn))
-            );
-        }),
-        tap(a => {
-          if (a.component) {
-            const list = a.component.map(
-              c => new CompTile(c.id, c.service_id, c.display_name, constraint(c.constraint), c.service_state !== 'created', getActions(c))
-            );
-
-            this.sourceMap.set('compo', [...this.sourceMap.get('compo'), ...list]);
-          }
-        }),
-        tap((a: { component: adcmComponent[]; host: Host[]; hc: Post[] }) => {
-          if (a.hc) {
-            this.statePost.update(a.hc);
-            this.loadPost.update(a.hc);
-            this.setRelations(a.hc);
-            this.checkConstraints();
-            this.initFlag = false;
-          }
+          if (a.hc) this.initFlag = false;
         }),
         this.takeUntil()
       )
-      .subscribe();
-  }
-
-  setRelations(a: Post[]) {
-    a.forEach(p => {
-      const host = this.sourceMap.get('host').find(h => h.id === p.host_id),
-        service = this.sourceMap.get('compo').find(s => s.id === p.component_id);
-      if (host && service) {
-        if (this.actionParameters) {
-          service.relations = [...service.relations, host];
-          const clone = { ...service };
-          clone.disabled = service.actions.every(k => k !== 'remove');
-          host.relations = [...host.relations, clone];
-        } else {
-          host.relations = [...host.relations, service];
-          service.relations = [...service.relations, host];
-        }
-      }
-    });
-  }
-
-  clearRelations() {
-    this.sourceMap.get('host').map(h => (h.relations = []));
-    this.sourceMap.get('compo').map(s => (s.relations = []));
-  }
-
-  checkConstraints() {
-    this.countConstraint = this.sourceMap.get('compo').reduce((a, c) => {
-      if (c.limit && c.limit[0]) {
-        if (c.limit[0] === '+' && c.relations.length < this.sourceMap.get('host').length) a++;
-        else if (c.limit[0] > c.relations.length) a++;
-      }
-      return a;
-    }, 0);
-  }
-
-  clearServiceFromHost(data: { rel: CompTile; model: HostTile }) {
-    this.clear([data.model, data.rel]);
-    this.statePost.delete(new Post(data.model.id, data.rel.service_id, data.rel.id));
-  }
-
-  clearHostFromService(data: { rel: HostTile; model: CompTile }) {
-    this.clear([data.rel, data.model]);
-    this.statePost.delete(new Post(data.rel.id, data.model.service_id, data.model.id));
-  }
-
-  selectHost(host: HostTile) {
-    this.stream.target = host;
-    this.getLink('compo')
-      .getSelected('host')
-      .fork(this.handleLink, this.handleSelect);
-  }
-
-  selectService(comp: CompTile) {
-    this.stream.target = comp;
-    this.getLink('host')
-      .getSelected('compo')
-      .fork(this.handleLink, this.handleSelect);
-  }
-
-  getLink(name: string) {
-    this.stream.linkSource = this.sourceMap.get(name);
-    this.stream.link = this.stream.linkSource.find(s => s.isSelected);
-    this.stream.linkSource.forEach(s => (s.isLink = false));
-    return this;
-  }
-
-  getSelected(name: string) {
-    this.stream.selected = this.sourceMap.get(name).find(_s => _s.isSelected);
-    if (this.stream.selected) this.stream.selected.isSelected = false;
-    return this;
-  }
-
-  fork(one: { (): void; (): void; call?: any }, two: { (): void; (): void; call?: any }) {
-    if (this.stream.link) one.call(this);
-    else if (this.stream.selected !== this.stream.target) two.call(this);
-  }
-
-  handleSelect() {
-    this.stream.target.isSelected = true;
-    this.stream.target.relations.forEach(e => (this.stream.linkSource.find(s => s.name === e.name && s.id === e.id).isLink = true));
-  }
-
-  checkActions(host: HostTile, com: CompTile, action: 'add' | 'remove'): boolean {
-    const flag = this.loadPost.data.some(a => a.component_id === com.id && a.service_id === com.service_id && a.host_id === host.id);
-    if (com.actions && com.actions.length) {
-      if (action === 'remove') {
-        if (flag) return com.actions.some(a => a === 'remove');
-        else return true;
-      }
-      if (action === 'add') {
-        if (flag) return true;
-        else return com.actions.some(a => a === 'add');
-      }
-    } else return true;
-  }
-
-  handleLink() {
-    const str = this.stream;
-    const isComp = this.stream.target instanceof CompTile;
-    const CurrentServiceComponent = (isComp ? this.stream.target : this.stream.link) as CompTile,
-      CurrentHost = isComp ? this.stream.link : this.stream.target;
-    const post = new Post(CurrentHost.id, CurrentServiceComponent.service_id, CurrentServiceComponent.id);
-
-    if (str.link.relations.find(e => e.id === str.target.id)) {
-      if (!this.checkActions(CurrentHost, CurrentServiceComponent, 'remove')) return;
-      this.clear([str.target, str.link]);
-      this.statePost.delete(post);
-    } else if (this.noLimit(CurrentServiceComponent)) {
-      if (!this.checkActions(CurrentHost, CurrentServiceComponent, 'add')) return;
-      str.link.relations.push(str.target);
-      str.target.relations.push(str.link);
-      str.target.isLink = true;
-      this.statePost.add(post);
-    }
-    this.checkConstraints();
-  }
-
-  clear(tiles: Tile[]) {
-    for (let a of tiles) {
-      const name = a instanceof HostTile ? 'host' : 'compo';
-      const link = this.sourceMap.get(name).find(h => h.id === a.id);
-      const rel = tiles.find(b => b !== a);
-      link.relations = link.relations.filter(r => r.id !== rel.id);
-      a.isLink = false;
-    }
-    this.checkConstraints();
-  }
-
-  save() {
-    const send = { cluster_id: this.cluster.id, hc: this.statePost.data };
-    this.saveFlag = true;
-    this.api
-      .post<Post[]>(this.cluster.hostcomponent, send)
-      .pipe(
-        take(1),
-        catchError(e => {
-          this.showPopup = false;
-          return throwError(e);
-        })
-      )
-      .subscribe(data => {
-        this.loadPost.update(data);
-        this.statePost.update(data);
-        this.saveResult.emit(data);
-        this.showPopup = true;
-        setTimeout(() => (this.showPopup = false), 2000);
-        this.saveFlag = false;
+      .subscribe(_ => {
+        this.serviceComponents = this.service.Components;
+        this.hosts = this.service.Hosts;
       });
   }
 
-  cancel() {
-    this.statePost.clear();
-    this.statePost.update(this.loadPost.data);
-    this.clearRelations();
-    this.setRelations(this.loadPost.data);
-    this.checkConstraints();
-    this.sourceMap.get('host').map(a => {
-      a.isSelected = false;
-      a.isLink = false;
-    });
-    this.sourceMap.get('compo').map(a => {
-      a.isSelected = false;
-      a.isLink = false;
+  clearRelations() {
+    this.service.clearAllRelations();
+  }
+
+  clearServiceFromHost(data: { rel: CompTile; model: HostTile }) {
+    this.service.clearServiceFromHost(data);
+  }
+
+  clearHostFromService(data: { rel: HostTile; model: CompTile }) {
+    this.service.clearHostFromService(data);
+  }
+
+  selectHost(host: HostTile) {
+    this.service.takeHost(host);
+  }
+
+  selectService(comp: CompTile) {
+    this.service.takeComponent(comp);
+  }
+
+  save() {
+    this.saveFlag = true;
+    this.service.saveSource(this.cluster).subscribe(data => {
+      this.saveResult.emit(data);
+      this.showPopup = true;
+      setTimeout(() => (this.showPopup = false), 2000);
+      this.saveFlag = false;
     });
   }
 
-  noLimit(comp: Tile) {
-    if (comp.limit && Array.isArray(comp.limit) && comp.limit.length) {
-      const a = comp.limit,
-        b = a.length - 1,
-        c = comp.relations.length;
-      // if (a.includes('odd')) {
-      //   //return comp.relations.length === 0 || comp.relations.length % 2 === 0;
-      // }
-      return a[b] === '+' || a[b] > c;
-    } else return true;
+  cancel() {
+    this.service.cancel();
   }
 }
