+++ web/src/app/shared/components/hostmap/take.service.spec.ts
@@ -1,0 +1,36 @@
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+import { HttpClientModule } from '@angular/common/http';
+import { TestBed } from '@angular/core/testing';
+import { ApiService } from '@app/core/api';
+import { StoreModule } from '@ngrx/store';
+
+import { TakeService } from './take.service';
+import { reducers } from '@app/core/store';
+
+describe('TakeService', () => {
+  beforeEach(() =>
+    TestBed.configureTestingModule({
+      imports: [HttpClientModule, StoreModule.forRoot(reducers)],
+      providers: [ApiService]
+    })
+  );
+
+  it('should be created', () => {
+    const service: TakeService = TestBed.get(TakeService);
+    expect(service).toBeTruthy();
+  });
+
+  it('should be init take', () => {
+    const service: TakeService = TestBed.get(TakeService);
+  });
+});
